double改为float

+++ app/src/main/java/hust/cc/asynchronousacousticlocalization/physical/SignalGenerator.java
@@ -12,8 +12,8 @@ public class SignalGenerator implements FlagVar {
      * @param f - fmin
      * @return audio samples in short format
      */
-    public static short[] upChirpGenerator(int fs, double t, int b, int f){
-        double[] x = chirpGenerator(fs, t, b, f, 0);
+    public static short[] upChirpGenerator(int fs, float t, int b, int f){
+        float[] x = chirpGenerator(fs, t, b, f, 0);
         waveformReshaping(x);
         short[] samples = new short[x.length];
         for(int i = 0; i < x.length; i++){
@@ -30,8 +30,8 @@ public class SignalGenerator implements FlagVar {
      * @param f - fmax
      * @return audio samples in short format
      */
-    public static short[] downChirpGenerator(int fs, double t, int b, int f){
-        double[] x = chirpGenerator(fs, t, b, f, 1);
+    public static short[] downChirpGenerator(int fs, float t, int b, int f){
+        float[] x = chirpGenerator(fs, t, b, f, 1);
         waveformReshaping(x);
         short[] samples = new short[x.length];
         for(int i = 0; i < x.length; i++){
@@ -47,18 +47,18 @@ public class SignalGenerator implements FlagVar {
      * @param b - bandwidth of the chirp signal
      * @param f - fmin for up chirp signal and fmax for the down chirp signal
      * @param type - 0 for up chirp signal and 1 for down chirp signal
-     * @return chirp samples in double format
+     * @return chirp samples in float format
      */
-    public static double[] chirpGenerator(int fs, double t, int b, int f, int type){
+    public static float[] chirpGenerator(int fs, float t, int b, int f, int type){
         int n = (int)(fs * t);
-        double[] samples = new double[n];
+        float[] samples = new float[n];
         if( type == 0 ) {
             for (int i = 0; i < n; i++) {
-                samples[i] = Math.cos(2 * Math.PI * f * i / fs + Math.PI * b * i * i / t / fs / fs);
+                samples[i] = (float)Math.cos(2 * Math.PI * f * i / fs + Math.PI * b * i * i / t / fs / fs);
             }
         }else{
             for (int i = 0; i < n; i++) {
-                samples[i] = Math.cos(2 * Math.PI * f * i / fs - Math.PI * b * i * i / t / fs / fs);
+                samples[i] = (float)Math.cos(2 * Math.PI * f * i / fs - Math.PI * b * i * i / t / fs / fs);
             }
         }
         return samples;
@@ -69,9 +69,9 @@ public class SignalGenerator implements FlagVar {
      * slowly ramping up the amplitude of the 100 samples and reversely perform it on the last 100 samples
      * @param samples
      */
-    public static void waveformReshaping(double samples[]){
+    public static void waveformReshaping(float samples[]){
         int k = 100;
-        double coefficients = 1.0 / k;
+        float coefficients = 1.0f / k;
 
         for(int i = 0; i < k; i++){
             samples[i] = samples[i] * (i + 1) / k;
