+++ app/src/main/java/hust/cc/asynchronousacousticlocalization/processing/IndexMaxVarInfo.java
@@ -1,7 +1,7 @@
 package hust.cc.asynchronousacousticlocalization.processing;
 
 public class IndexMaxVarInfo {
-    public double maxVar;
+    public float maxVar;
     public int index;
     public boolean isReferenceSignalExist = false;
 }
